changed dynamic bounding spheres to reassign their center-x based on the bounding box.
this is not ideal but is going to improve the player experiance quite a bit and can be removed when the major ship overhaul is done. I sent an email to Kasper regarding making some of these ships have static bounding spheres instead.

This might need a big QA check on ships with siege mode to make sure they don't get off-centered a lot. (especially those that are asymmetrical) 

https://jira.ccpgames.com/browse/TQ-131757

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/Trinity/": change = 1572977]