Add skip condition to tests using WithValidRenderContext requiring >0 device
If a test which uses WithValidRenderContext is run on a machine that doesn't have
a device then the test is skipped. This is required to get tests passing on TeamCity
which doesn't have GPU agents.

Also removed previous experiment to buffer test which should be covered by this new change.